% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{abundance_match_census_habitat}
\alias{abundance_match_census_habitat}
\title{Abundance per quadrat, ensuring consistent \code{gridsize} and \code{plotdim}.}
\usage{
abundance_match_census_habitat(censdata, habitats)
}
\arguments{
\item{censdata}{A table of plot census data.}

\item{habitats}{Data frame giving the habitat designation for each 20x20
quadrat.}
}
\value{
A dataframe giving the abundance per quadrat.
}
\description{
Abundance per quadrat, ensuring consistent \code{gridsize} and \code{plotdim}. This
funciton doe not require the user to provide plot dimensionsbecause it
extracts \code{gridsize} and \code{plotdim} from habitat data.
}
\section{Acknowledgement}{

Thanks to Gabriel Arellano for suggesting that the match plot dimensions
between habitat data and census data should be enforced programatically.
}

\examples{
abund <- abundance_match_census_habitat(bci::bci12full1, bci::bci_habitat)
abund[1:3, 1:10]
}
\seealso{
\code{\link[=extract_gridsize]{extract_gridsize()}}, \code{\link[=extract_plotdim]{extract_plotdim()}},
\code{\link[forestr:abundanceperquad]{forestr::abundanceperquad()}}.
}
