% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_.R
\name{abundance_n2}
\alias{abundance_n2}
\title{Count individuals.}
\usage{
abundance_n2(x, group_by = c("sp", "status"), only_alive = TRUE, ...)
}
\arguments{
\item{x}{Census.}

\item{group_by}{Sting giving the variables to group by.}

\item{only_alive}{If TRUE, filter only alive individuals.}

\item{...}{Arguments passed to \code{\link[base:tapply]{base::tapply()}}.}
}
\value{
Count of individuals in each group.
}
\description{
Count individuals.
}
\examples{
stem <- forestr::bci12s6mini
head(abundance_n2(stem))
head(abundance_n2(stem, group_by = c("status"), only_alive = FALSE))
head(abundance_n2(stem, group_by = c("status"), only_alive = FALSE))
}
\seealso{
\code{\link[base:tapply]{base::tapply()}}.
}
