% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_.R
\name{abundance_n_d_plyr}
\alias{abundance_n_d_plyr}
\alias{abundance_n_dplyr}
\alias{abundance_n_dplyr_se}
\alias{abundance_n_dplyr_se_}
\alias{abundance_n_plyr}
\title{Abundance with dplyr and plyr (deprecated)}
\usage{
abundance_n_dplyr(x, ..., only_alive = TRUE)

abundance_n_dplyr_se(x, groups = c("sp", "status"), only_alive = TRUE)

abundance_n_dplyr_se_(x, groups = c("sp", "status"), only_alive = TRUE)

abundance_n_plyr(x, groups = c("sp", "status"), alive_only = TRUE)
}
\arguments{
\item{x}{Census.}

\item{...}{Bare names of variables to group by.}

\item{only_alive}{If TRUE, filter only alive individuals.}

\item{groups}{Character string giving the names of variables to group by.}
}
\value{
Tally if groups = NULL, else, a count by group.
}
\description{
\itemize{
\item Abundance with dplyr, grouping with bare variable names.
\itemize{
\item \code{\link[=abundance_n_dplyr]{abundance_n_dplyr()}} uses the latest "tidy eval" approach.
}
\item Abundance with dplyr, grouping with a character string
\itemize{
\item \code{\link[=abundance_n_dplyr_se]{abundance_n_dplyr_se()}} uses latest tidy eval approach.
\item \code{\link[=abundance_n_dplyr_se_]{abundance_n_dplyr_se_()}} uses deprecated approach.
\item \code{\link[=abundance_n_plyr]{abundance_n_plyr()}} uses plyr (deprecated).
}
}
}
\examples{
stem <- bci12s7mini

# Using plyr, the deprecated ancestor of dplyr.
head(abundance_n_plyr(stem))
head(abundance_n_plyr(stem), groups = c("quadrat", "sp"), alive_only = FALSE)

# Using non-standard evaluation.
abundance_n_dplyr(stem)
abundance_n_dplyr(stem, sp, only_alive = FALSE)

# Standard evaluation using the tidy eval approach.
abundance_n_dplyr_se(stem)
abundance_n_dplyr_se(stem, groups = NULL)

# Standard evaluation using older approach. See ?group_by_ (ends in "_").
abundance_n_dplyr_se_(stem)
abundance_n_dplyr_se_(stem, groups = c("status"))
abundance_n_dplyr_se_(stem, groups = NULL)

}
