% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{assert_are_names_matching}
\alias{assert_are_names_matching}
\title{Assert important names match names in data.}
\usage{
assert_are_names_matching(.data, match, ...)
}
\arguments{
\item{.data}{Data}

\item{match}{Names to match with names of \code{.data}.}

\item{...}{Arguments passed to \code{\link[assertive:assert_all_are_matching_regex]{assertive::assert_all_are_matching_regex()}},
e.g. \code{severity = "stop"} (default), \code{severity = "warning"}.}
}
\value{
Nothing, warning, error, etc. See
\code{\link[assertive:assert_all_are_matching_regex]{assertive::assert_all_are_matching_regex()}}.
}
\description{
Assert important names match names in data.
}
\examples{
# Passes silently
assert_are_names_matching(data.frame(a = 1), match = c("a"))
# Errs
testthat::expect_error(
  assert_are_names_matching(data.frame(a = 1), match = c("aa"))
)
# Warns
assert_are_names_matching(data.frame(a = 1), match = c("aa"),
  severity = "warning")
}
\keyword{internal}
