% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basal_area.R
\name{basal_area}
\alias{basal_area}
\alias{basal_area.numeric}
\title{Basal area of each individual.}
\usage{
basal_area(x)

\method{basal_area}{numeric}(x, group_by, only_alive)
}
\arguments{
\item{x}{A vector giving the individuals' diameter or a dataframe with
individuals' diameter stored in the column \code{dbh}.}

\item{group_by}{Character string giving the names of variables to group by.}

\item{only_alive}{If TRUE, counts only alive individuals (\code{status == "A"}).
If FALSE counts individuals of all available statuses.}
}
\value{
Each individual calculation is simply the area of a circle given its
diameter; thus the unit of the returned value is the square of the input
units unit (i.e. for input in mm, output is mm^2; for input in cm, output is
cm^2, and so on).
\itemize{
\item \code{basal_area.numeric()}: A numeric vector giving the basal area of each
individual.
}
}
\description{
Basal area of each individual. The input can be either a vector or a
dataframe.
}
\examples{
stem <- fgeo.data::luquillo_stem_random_tiny

head(stem$dbh)
head(basal_area.numeric(stem$dbh))

head(basal_area(stem))

# Silent
head(suppressMessages(basal_area(stem)))
}
\seealso{
Other grouped summaries: \code{\link{abundance}}
}
