% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bci_habitat}
\alias{bci_habitat}
\title{Habitat data from Barro Colorado Island.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1250 rows and 4 columns.}
\usage{
bci_habitat
}
\description{
Habitat data from Barro Colorado Island.
}
\seealso{
https://forestgeo.github.io/bci/reference/bci_habitat.html
}
\keyword{datasets}
