% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bci_mini}
\alias{bci_mini}
\title{Minimal data from Barro Colorado Island for examples.}
\format{An object of class \code{data.frame} with 10000 rows and 20 columns.}
\usage{
bci_mini
}
\description{
Minimal data from Barro Colorado Island for examples.
}
\seealso{
https://forestgeo.github.io/bci
}
\keyword{datasets}
