% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byyr.R
\name{byyr}
\alias{byyr}
\alias{byyr_abundance}
\alias{byyr_basal_area}
\title{Create tables of abundance and basal area by (round mean) year.}
\usage{
byyr_abundance(vft)

byyr_basal_area(vft)
}
\arguments{
\item{vft}{A dataframe; particularly a ForestGEO ViewFullTable.}
}
\value{
A dataframe.
}
\description{
Create tables of abundance and basal area by (round mean) year.
}
\examples{
# Or load all fgeo packages in one step with library(fgeo)
library(fgeo.tool)
library(fgeo.base)
  
vft <- tibble(
  PlotName = c("luq", "luq", "luq", "luq", "luq", "luq", "luq", "luq"),
  CensusID = c(1L, 1L, 1L, 1L, 2L, 2L, 2L, 2L),
  TreeID = c(1L, 1L, 2L, 2L, 1L, 1L, 2L, 2L),
  StemID = c(1.1, 1.2, 2.1, 2.2, 1.1, 1.2, 2.1, 2.2),
  Status = c("alive", "dead", "alive", "alive", "alive", "gone",
    "dead", "dead"),
  DBH = c(1L, NA, 2L, 3L, 2L, NA, 3L, 4L),
  Genus = c("Gn", "Gn", "Gn", "Gn", "Gn", "Gn", "Gn", "Gn"),
  SpeciesName = c("spp", "spp", "spp", "spp", "spp", "spp", "spp", "spp"),
  ExactDate = c("2001-01-01", "2001-01-01", "2001-01-01", "2001-01-01",
    "2002-01-01", "2002-01-01", "2002-01-01",
    "2002-01-01"),
  PlotCensusNumber = c(1L, 1L, 1L, 1L, 2L, 2L, 2L, 2L),
  Family = c("f", "f", "f", "f", "f", "f", "f", "f"),
  Tag = c(1L, 1L, 2L, 2L, 1L, 1L, 2L, 2L)
)
vft

# All trees are of the same species. There are two trees, each with two stems.
# In census 1, the count of alive trees should be 2 because both trees are alive,
#   but note that one stem is dead (StemID = 1.2).
# In census 2 the count of alive trees should be 1:
#   * One tree is alive (TreeID = 1) although one stem is gone (StemID = 1.2);
#   * One tree is dead (TreeID = 2) because both its stems are dead.

vft <- vft \%>\% 
  # Determine tree status based on stem status.
  fgeo.tool::add_status_tree(status_a = "alive", status_d = "dead") \%>\%
  # Remove dead trees.
  fgeo.tool::drop_dead_tree(.status = "dead")

byyr_abundance(vft)

ba <- byyr_basal_area(vft)

# Convert units and standardize by plot size in hectares
years <- c("2001", "2002")
in_he <- convert_unit_at(ba, .at = years, from = "mm2", to = "hectare")
standardize_at(in_he, .at = years, denominator = 50)
}
