% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality.R
\name{calcMortIndivTable}
\alias{calcMortIndivTable}
\title{Internal.}
\usage{
calcMortIndivTable(mtable, by = "species")
}
\description{
Internal.
}
\seealso{
Other functions from http://ctfs.si.edu/Public/CTFSRPackage/: \code{\link{demography}},
  \code{\link{drp}}, \code{\link{fill.dimension}},
  \code{\link{find.climits}},
  \code{\link{mortality.calculation}},
  \code{\link{mortality.dbh}},
  \code{\link{mortality.eachspp}}
}
\keyword{internal}
