% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_distinct.R
\name{count_distinct}
\alias{count_distinct}
\title{Count distinct values of any variable (optionally by groups).}
\usage{
count_distinct(.data, .var)
}
\arguments{
\item{.data}{A dataframe, commonly grouped with \code{group_by()}.}

\item{.var}{A variable to count distinct occurrences.}
}
\value{
An object of the same class as .data. One grouping level will be
dropped.
}
\description{
This function works with grouped data (via
\code{dplyr::group_by()}). The output will have one row for each group.
}
\examples{
library(dplyr)

# Compare to dplyr::summarize() and dplyr::n_distinct()
library(dplyr)

by_cyl <- group_by(mtcars, cyl)
count_distinct(by_cyl, mpg)
# Same
summarize(by_cyl, n = n_distinct(mpg))

n_distinct(mtcars)
# Same
nrow(mtcars)

\dontrun{
# Fails; you must supply `.var` to count distinct values if that variable
count_distinct(mtcars)
}

count_distinct(mtcars, cyl)
# Same
n_distinct(mtcars$cyl)
# Same
length(unique(mtcars$cyl))

\dontrun{
# Fails because it expect not column names but additional vectors (via `...`)
dplyr::n_distinct(mtcars, cyl)
}
}
\seealso{
\code{\link[dplyr:group_by]{dplyr::group_by()}}, \code{\link[dplyr:summarise]{dplyr::summarise()}}, \code{\link[=drop_dead_tree]{drop_dead_tree()}},
\code{\link[=drop_dead_stem]{drop_dead_stem()}}.
}
\concept{abundance functions.}
\concept{functions not specific to ForestGEO}
