% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_alive.R
\name{count_status_by}
\alias{count_status_by}
\title{Count individuals of a given status by groups.}
\usage{
count_status_by(x, .status, ...)
}
\arguments{
\item{x}{Census.}

\item{.status}{A character vector giving the statuses to keep.}

\item{...}{Optional. Bare name of variables to group by (passed to
\code{\link[dplyr:count]{dplyr::count()}}.}
}
\value{
A count summary by group.
}
\description{
Count individuals of a given status by groups.
}
\keyword{internal}
