% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_woods.R
\name{count_woods}
\alias{count_woods}
\alias{abundance}
\alias{count_trees}
\alias{count_saplings}
\alias{count_saplings_and_trees}
\title{Count unique trees based on the dbh of their largest stem.}
\usage{
count_woods(.data, ...)

count_trees(.data)

count_saplings(.data)

count_saplings_and_trees(.data)
}
\arguments{
\item{.data}{A dataframe; particularly a ForestGEO census or ViewFullTable.}

\item{...}{Expressions to pick stems of specific \code{dbh} -- where \emph{stems}
refers to the largest stem of each tree.}
}
\value{
A dataframe with preserved groups (if any).
}
\description{
\code{count_woods()} is a general function that takes any dataframe and any number
of expressions to filter that dataframe. The other functions are shortcuts:
\itemize{
\item \code{count_trees()} picks stems of 100 mm dbh and above.
\item \code{count_saplings()} picks stems between 10 mm dbh inclusive and 100 mm dbh
exclusive.
\item \code{count_saplings_and_trees()} picks stems of 10 mm dbh and above.
}
}
\details{
This functions count unique treeid by censusid (automatically) and any number
of additional grouping variables (via \code{dplyr::group_by()}. Data with
multi-stem trees is first collapsed by picking the stem (of each treeid) with
maximum dbh.
}
\examples{
library(dplyr)

census <- tibble::tribble(
    ~sp, ~treeID, ~stemID, ~hom, ~dbh,
  "sp1",     "1",   "1.1",  130,   10,
  "sp1",     "1",   "1.2",  130,  100,
  "sp2",     "2",   "2.1",  130,   22,
  "sp2",     "2",   "2.2",  130,   99,
  "sp2",     "2",   "2.3",  130,   NA
)

# Count all woods
count_woods(census)

by_sp_treeid <- group_by(census, sp, treeID)
count_trees(by_sp_treeid)

by_sp <- group_by(census, sp)
count_woods(by_sp)

count_woods(by_sp, dbh >= 100)
# Same
count_trees(by_sp)

count_woods(by_sp, dbh > 10, dbh < 100)
# Same
count_saplings(by_sp)

count_woods(by_sp, dbh >= 10)
# Same
count_saplings_and_trees(by_sp)
}
\seealso{
Other functions for fgeo census and vft.: \code{\link{count_distinct_treeid}}
}
\concept{functions for fgeo census and vft.}
