% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_abundance.R
\name{difference_among_grid_steps}
\alias{difference_among_grid_steps}
\title{From x and y columns of habitat data, get difference between grid steps.}
\usage{
difference_among_grid_steps(habitat_x_or_y)
}
\arguments{
\item{habitat_x_or_y}{Column x or y of habitat data, e.g. bci::bci_habitat$x.}
}
\value{
A non negative scalar
}
\description{
From x and y columns of habitat data, get difference between grid steps.
}
