% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_dead_trees_by_cns.R
\name{drop_dead_trees_by_cns}
\alias{drop_dead_trees_by_cns}
\title{Drop dead trees by census from a ViewFullTable.}
\usage{
drop_dead_trees_by_cns(vft, status_d = "dead", .valid_status = c("dead",
  "alive", "broken below", "missing"))
}
\arguments{
\item{vft}{A dataframe; particularly a ForestGEO ViewFullTable.}

\item{status_d}{String indicating dead stems in the column \code{Status} of \code{vft}
(commonly "dead" or "D").}

\item{.valid_status}{String giving possible values of \code{Status}.}
}
\value{
A dataframe.
}
\description{
If you have a ViewFullTable and want to compute some summary on exclusively
alive trees, you can't just drop dead trees across the entire ViewFullTable.
If you do that you'll loose trees (tags) which were alive in earlier
censuses. This function helps you solve that problem by (1) grouping the
data by census; (2) calculating the status of each tree based on the status
of its stem(s); (3) dropping dead trees (i.e. trees which all stems are
dead).
}
\examples{
vft <- data.frame(
  TreeID = c("0001", "0001", "0002", "0002"),
  PlotName = "p",
  Status = c("dead", "dead", "alive", "alive"),
  DBH = c(NA, NA, 10, 100),
  ExactDate = c("2000-01-01", "2001-01-01", "2000-01-01", "2001-01-01"),
  PlotCensusNumber = c(1, 2, 1, 2),
  CensusID = c(1, 2, 1, 2),
  Genus = c("A", "A", "B", "B"),
  SpeciesName = c("a", "a", "b", "b"),
  Family = "f",
  stringsAsFactors = FALSE
)
vft

drop_dead_trees_by_cns(vft)
}
\concept{functions by census.}
\concept{functions for fgeo vft.}
