% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_if_na.R
\name{drop_if_na}
\alias{drop_if_na}
\title{Drop if missing values.}
\usage{
drop_if_na(dfm, x)
}
\arguments{
\item{dfm}{A dataframe.}

\item{x}{String giving a column name of \code{dfm}.}
}
\description{
Valuable mostly for its warning.
}
