% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_insensitive}
\alias{extract_insensitive}
\alias{detect_insensitive}
\title{Detect and extract matching strings ignoring case.}
\usage{
extract_insensitive(x, y)

detect_insensitive(x, y)
}
\arguments{
\item{x, y}{A string.}
}
\value{
\code{detect_*} and \code{extract_*} return a logical vector and a string.
}
\description{
Detect and extract matching strings ignoring case.
}
\examples{
detect_insensitive(c("StemID", "treeID"), c("stemid", "n"))
extract_insensitive(c("StemID", "treeID"), c("stemid", "n"))

vft <- data.frame(TreeID = 1, Status = 1)
extract_insensitive(names(vft), tolower(names(vft)))
extract_insensitive(tolower(names(vft)), names(vft))
}
\keyword{internal}
