% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.abundance-package.r
\docType{package}
\name{fgeo.abundance-package}
\alias{fgeo.abundance}
\alias{fgeo.abundance-package}
\title{fgeo.abundance: R packages for the Analysis of Forest Dynamics}
\description{
Tools for the analysis of forest dynamics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/forestgeo/fgeo.abundance}
  \item Report bugs at \url{https://github.com/forestgeo/fgeo.abundance/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{leporem@si.edu} [contractor]

Other contributors:
\itemize{
  \item CTFS-ForestGEO \email{ForestGEO@si.edu} [copyright holder, funder]
}

}
\keyword{internal}
