% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_abundance.R
\name{fgeo_abundance}
\alias{fgeo_abundance}
\alias{fgeo_basal_area}
\title{Create tables fgeo_abundance and fgeo_basal_area.}
\usage{
fgeo_abundance(vft, .status = "dead", exclude = TRUE,
  .valid_status = c("dead", "alive", "broken below", "missing"))

fgeo_basal_area(vft, .status = "dead", exclude = TRUE,
  .valid_status = c("dead", "alive", "broken below", "missing"))
}
\arguments{
\item{vft}{Dataframe; particularly a ForestGEO ViewFullTable.}

\item{.status}{Character vector; Must be one of possible values of the
variable giving the status of the dataframe \code{x}.}

\item{exclude}{Logical; \code{TRUE} filters the data to exclude the values passed
to \code{.status}.}

\item{.valid_status}{String giving possible values of \code{Status}.}
}
\value{
A dataframe.
}
\description{
Create tables fgeo_abundance and fgeo_basal_area.
}
