% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{fill.dimension}
\alias{fill.dimension}
\title{Internal.}
\usage{
fill.dimension(dataarray, class1, class2, fill = 0)
}
\description{
Internal.
}
\seealso{
Other functions from http://ctfs.si.edu/Public/CTFSRPackage/: \code{\link{calcMortIndivTable}},
  \code{\link{drp}}, \code{\link{find.climits}},
  \code{\link{mortality.calculation}},
  \code{\link{mortality.dbh}},
  \code{\link{mortality.eachspp}}, \code{\link{mortality}}
}
\keyword{internal}
