% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_utilities.R
\name{fill_dimension}
\alias{fill_dimension}
\alias{fill.dimension}
\title{Fill 2 dimensional data with custom value.}
\usage{
fill_dimension(.data, class1, class2, fill = 0)
}
\arguments{
\item{.data}{A two-dimension array, matrix or data.frame.}

\item{class1, class2}{A vector of the same length  of \code{.data} or longer.}

\item{fill}{A numeric value to fill missing values with.}
}
\description{
This function fills out an array of 2 dimensions with custom values for extra
columns and rows as named in \code{class1} and \code{class2}. If a column or row is
missing, it will be filled with the value given by fill. It is useful for
results of table or tapply when some elements had no records.
}
\examples{
class1 <- 1:2
class2 <- letters[1:5]
fill = 0
.data <- array(
  data = c(1, 2, NA),
  dim = c(2, 4),
  dimnames = list(1:2, letters[1:4])
)
}
