% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_na_with_single_unique.R
\name{fill_na_with_single_unique}
\alias{fill_na_with_single_unique}
\title{Replace missing values with single occurrences of a vector or die trying.}
\usage{
fill_na_with_single_unique(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The input vector with missing values replaced by unique ocurrences
of x.
}
\description{
Replace missing values with single occurrences of a vector or die trying.
}
\examples{

fill_na_with_single_unique(c(1, NA))
fill_na_with_single_unique(c("a", NA))
fill_na_with_single_unique(c(TRUE, NA))

# Fails
# fill_na_with_single_unique(c(1, 2, NA))
}
\seealso{
\code{rlang::}%|%`.
}
