% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_alive.R
\name{filter_status}
\alias{filter_status}
\title{Filter individuals of a given status.}
\usage{
filter_status(x, .status)
}
\arguments{
\item{x}{Census.}

\item{.status}{A character vector giving the statuses to keep.}
}
\value{
Returns \code{x} filtered by \code{.status}.
}
\description{
Filter individuals of a given status.
}
