% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestr-package.R
\docType{package}
\name{forestr-package}
\alias{forestr}
\alias{forestr-package}
\title{forestr: R packages for the Analysis of Forest Dynamics}
\description{
Tools for the analysis of forest dynamics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/forestgeo/forestr}
  \item Report bugs at \url{https://github.com/forestgeo/forestr/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{leporem@si.edu} [contractor]

Other contributors:
\itemize{
  \item CTFS-ForestGEO \email{ForestGEO@si.edu} [copyright holder, funder]
}

}
\keyword{internal}
