% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestr-package.r
\docType{package}
\name{forestr}
\alias{forestr}
\alias{forestr-package}
\alias{forestr-package}
\title{forestr: tools for the analysis of forest dynamics.}
\description{
\strong{forestr} provides tools for the analysis of forest dynamics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/forestgeo/forestr}
  \item Report bugs at \url{https://github.com/forestgeo/forestr/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{leporem@si.edu} [contractor]

Other contributors:
\itemize{
  \item CTFS-ForestGEO \email{ForestGEO@si.edu} [copyright holder, funder]
}

}
