% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_summaries.R
\name{group}
\alias{group}
\title{A version of \code{\link[dplyr:group_by]{dplyr::group_by()}} that inputs "quoted" variable names.}
\usage{
group(x, group_by, only_alive)
}
\arguments{
\item{x}{Census data.}

\item{group_by}{Character string giving the names of variables to group by.}

\item{only_alive}{If TRUE, counts only alive individuals (`status == "A"`).
If FALSE counts individuals of all available statuses.}
}
\value{
A grouped tibble.
}
\description{
A version of \code{\link[dplyr:group_by]{dplyr::group_by()}} that inputs "quoted" variable names.
}
\examples{
df <- bciex::bci12s7mini
group(df, "sp", only_alive = TRUE)
group(df, c("quadrat", "sp"), only_alive = TRUE)
}
\seealso{
\code{\link[dplyr:group_by]{dplyr::group_by()}}.
}
