% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insensitive}
\alias{insensitive}
\title{Create a vector or dataframe with case insensitive name matching.}
\usage{
insensitive(x)
}
\arguments{
\item{x}{vector or dataframe to modify.}
}
\description{
Create a vector or dataframe with case insensitive name matching.
}
\section{Acknowledgment}{

Thanks to Neil Richardson for recommending this function
(https://twitter.com/enpiar).
}

\examples{
x <- c("abc" = 1, "def" = 2)
x["ABC"]
y <- insensitive(x)
y["ABC"]
y[["ABC"]]
y$ABC

vft <- data.frame(TreeID = 1)
insensitive(vft)[["TreeID"]]
insensitive(vft)[["treeid"]]
insensitive(vft)["TreeID"]
insensitive(vft)["treeid"]
insensitive(vft)$TreeID
insensitive(vft)$treeid

# Works
transform(insensitive(vft), new = treeid)
# dplyr::mutate(insensitive(vft), new = treeid)
# Fails
# transform(insensitive(vft), new = TreeID)
# dplyr::mutate(insensitive(vft), new = TreeID)
}
\seealso{
\code{httr::insensitive()}.
}
\author{
Hadley Wickham (see \code{httr::insensitive()}).
(https://twitter.com/hadleywickham)
}
\keyword{internal}
