% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality.R
\name{mortality.calculation}
\alias{mortality.calculation}
\title{Internal.}
\usage{
mortality.calculation(N, S, meantime)
}
\description{
Internal.
}
\seealso{
Other functions from http://ctfs.si.edu/Public/CTFSRPackage/: \code{\link{calcMortIndivTable}},
  \code{\link{demography}}, \code{\link{drp}},
  \code{\link{fill.dimension}}, \code{\link{find.climits}},
  \code{\link{mortality.dbh}},
  \code{\link{mortality.eachspp}}
}
\keyword{internal}
