% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality.R
\name{mortality.dbh}
\alias{mortality.dbh}
\title{Internal.}
\usage{
mortality.dbh(census1, census2, classbreak = c(10, 100, 300),
  alivecode = c("A", "AB", "AS"))
}
\description{
Internal.
}
\seealso{
Other functions from http://ctfs.si.edu/Public/CTFSRPackage/: \code{\link{calcMortIndivTable}},
  \code{\link{drp}}, \code{\link{fill.dimension}},
  \code{\link{find.climits}},
  \code{\link{mortality.calculation}},
  \code{\link{mortality.eachspp}}, \code{\link{mortality}}
}
\keyword{internal}
