% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor.R
\name{neighbor_count}
\alias{neighbor_count}
\alias{neighbor_basal_area}
\title{Count and basal area of neighbouring stems.}
\usage{
neighbor_count(.data, .subset = NULL, r, plotdim)

neighbor_basal_area(.data, .subset = NULL, r, plotdim)
}
\arguments{
\item{.data}{A Dataframe; particularly  a ForestGEO census table (tree or
stem).}

\item{.subset}{An optional dataframe giving the (\code{gx}, \code{gy}) coordinates of
specific individuals (\code{tag} and \code{sp}) for which to count neighbors. For
example, \code{.subset} may be a subset of only one species; or seedling that
are not part of census \code{.data}. \code{.subset} must have columns \code{gx}, \code{gy},
\code{sp}, and \code{tag}.}

\item{r}{Radius.}

\item{plotdim}{The x and y dimensions of the plot.}
}
\value{
A dataframe.
}
\description{
Includes all values of status. You shouild pick the values you want before
using these functions (e.g. to drop dead stems).
}
\examples{
tree <- fgeo.data::luquillo_tree5_random
neighbor_count(tree, r = 20, plotdim = c(1000, 500))
neighbor_basal_area(tree, r = 20, plotdim = c(1000, 500))
}
