% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nms_restore_matching}
\alias{nms_restore_matching}
\title{Rename an object based on case-insensitive match of the names of a reference.}
\usage{
nms_restore_matching(new, ref)
}
\arguments{
\item{new}{New object which names to restored if they match the reference.}

\item{ref}{Named object to use as reference.}
}
\value{
The output is \code{new} with as many names changed as case-insensitive
matches there are with the reference.
}
\description{
Rename an object based on case-insensitive match of the names of a reference.
}
\examples{
ref <- data.frame(COLUMN = 1, other = 1)
new <- data.frame(Column = 5, Other = 1, n = 5)
nms_restore_matching(new, ref)
}
