% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_plotname.R
\name{pick_plotname}
\alias{pick_plotname}
\title{Pick the given or first census (alpha sorted) from a ViewFullTable.}
\usage{
pick_plotname(vft, plot_nm = NULL)
}
\arguments{
\item{vft}{Dataframe; particularly a ForestGEO ViewFullTable.}

\item{plot_nm}{Length-1 character vector of the value of \code{PlotName} to pick
from \code{vft}.}
}
\value{
A dataframe.
}
\description{
This function performs a very simple action but is valuable mostly for its
checks.
}
\examples{
vft <- data.frame(PlotName = c("a", "b"), stringsAsFactors = FALSE)
pick_plotname(vft)
pick_plotname(vft, "b")
}
