% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev-remove_pdf.R
\name{rm_files}
\alias{rm_files}
\title{Remove files from base directory.}
\usage{
rm_files(base, file = "pdf$")
}
\arguments{
\item{base}{The base of a directory, e.g. "./vignettes/".}

\item{file}{A file name, e.g. myplot.pdf.}
}
\description{
Useful to clean .pdf and .png files generated by tests, vignettes, and
interactive use.
}
