% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smry_diversity.R
\name{smry_diversity}
\alias{smry_diversity}
\title{Shannon and Simpson diversity indices.}
\usage{
smry_diversity(x, col, index = c("shannon", "invsimpson", "simpson"))
}
\arguments{
\item{x}{A dataframe.}

\item{col}{A column in \code{x}.}

\item{index}{An index as described}
}
\value{
A summary of the input dataframe, with the additional column
\code{diversity}.
}
\description{
Adds a column \code{diversity}, of Shannon and Simpson diversity indices
calculated with \code{\link[vegan:diversity]{vegan::diversity()}}. The output is a dataframe, convenient
for workflows with general purpose tools such as \strong{dplyr} and \strong{ggplot2}.
}
\examples{
library(dplyr)
library(ggplot2)

cns <- data.frame(
  CensusID = factor(rep(c(1, 2), 3)),
  quadrat = "0000",
  sp = rep(paste0("sp", 1:3), 2),
  n = sample.int(100, 6)
)
cns

smry_diversity(cns, n)

# The output of `smry_diversity` flows well into common pipelines:

diversity <- cns \%>\%
  group_by(CensusID) \%>\%
  smry_diversity(n)
# Same
diversity <- group_by(cns, CensusID)
diversity <- smry_diversity(diversity, n)

diversity

# A plot
diversity \%>\%
  ggplot(aes(CensusID, value)) +
  geom_col(aes(fill = diversity), position = "dodge")
# A summary
diversity \%>\%
  group_by(diversity) \%>\%
  summarise(mean_diversity = mean(value))
}
\seealso{
\code{\link[vegan:diversity]{vegan::diversity()}}.
}
