% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{sp_names_by_n}
\alias{sp_names_by_n}
\alias{sp_alive_n}
\alias{sp_abund_n}
\title{Name of species which count is at or above a threshold.}
\usage{
sp_alive_n(censdata, n)

sp_abund_n(abundances, n)
}
\arguments{
\item{censdata}{A table of plot census data.}

\item{n}{A round number representing a count.}

\item{abundances}{The `abund` element of the list that results from
[forestr::abundanceperquad()].}
}
\value{
A character string giving the names of species which count of alive
stems (\code{\link[=sp_alive_n]{sp_alive_n()}}) or total abundance per quadrat (\code{\link[=sp_abund_n]{sp_abund_n()}}) is
at or above a threshold.
}
\description{
Name of species which count is at or above a threshold.
}
\examples{
# With census data
sp_alive_n(censdata = bci_mini, n = 1500)

# With abundance data
abund <- abundance_match_census_habitat(bci_mini, bci_habitat)
sp_abund_n(abund, n = 1500)
}
