% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_alive.R
\name{tally_status}
\alias{tally_status}
\title{Tally individuals of a given status.}
\usage{
tally_status(x, .status)
}
\arguments{
\item{x}{Census.}

\item{.status}{A character vector giving the statuses to keep.}
}
\value{
An integer giving the total number (tally) of individuals.
}
\description{
Tally individuals of a given status.
}
