% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{to_id}
\alias{to_id}
\alias{rowcol.to.index}
\alias{gxgy.to.index}
\alias{to_id_rowcol}
\alias{to_id_gxgy}
\title{Convert to quadrat indices.}
\usage{
to_id_rowcol(.row, .col, gridsize, plotdim)

to_id_gxgy(gx, gy, gridsize, plotdim)
}
\arguments{
\item{.row, .col}{Row and column number.}

\item{gridsize}{Side of the square quadrate.}

\item{plotdim}{The x and y dimensions of the plot.}

\item{gx, gy}{Tree x and y coordinate.}
}
\value{
A numeric vector of indices.
}
\description{
These functions output quadrat indices from either row and column numbers
(\link{to_id_rowcol}), or gx and gy coordinates (\link{to_id_gxgy}]). They replace
\link[ctfs:rowcol.to.index]{ctfs::rowcol.to.index} and \link[ctfs:gxgy.to.index]{ctfs::gxgy.to.index}, but these old names are
kept as aliases to help users find the newer names.
}
\examples{
library(dplyr)
mini_cns <- bci::bci12full7 \%>\% 
  select(gx, gy) \%>\% 
  sample_n(10)
to_id_gxgy(mini_cns$gx, mini_cns$gy, gridsize = 20, plotdim = c(1000, 500))

to_id_rowcol(.row = 1:10, .col = 6:15, gridsize = 20, plotdim = c(1000, 500))
}
\author{
Richard Condit.
}
