% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{to_id_rowcol}
\alias{to_id_rowcol}
\alias{rowcol.to.index}
\title{Takes row and column numbers and identifies the quadrate number (index).}
\usage{
to_id_rowcol(.row, .col, gridsize, plotdim)
}
\arguments{
\item{.row}{Row number.}

\item{.col}{Column number.}

\item{gridsize}{Side of the square quadrate.}

\item{plotdim}{The x and y dimensions of the plot.}
}
\value{
A numeric vector of indices.
}
\description{
From row and column numbers (based on \code{gridsize}) outputs the corresponding
quadrat number, i.e. index (id).
}
\examples{
to_id_rowcol(.row = 1:10, .col = 1:10, gridsize = 20, plotdim = c(1000, 500))
}
