% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgn_diversity.R
\name{vgn_diversity}
\alias{vgn_diversity}
\title{Shannon and Simpson diversity indices.}
\usage{
vgn_diversity(x, abundance, index = c("specnumber", "shannon", "invsimpson",
  "simpson"))
}
\arguments{
\item{x}{A dataframe.}

\item{abundance}{A numeric column in \code{x} giving the abundance by species.}

\item{index}{An index as described}
}
\value{
A summary of the input dataframe, with the additional column
\code{diversity}.
}
\description{
Adds a column \code{diversity}, of Shannon and Simpson diversity indices
calculated with \code{\link[vegan:diversity]{vegan::diversity()}}. The output is a dataframe, convenient
for workflows with general purpose tools such as \strong{dplyr} and \strong{ggplot2}.
}
\section{Acknowledgements}{

\itemize{
\item David Kenfack inspired the need for this function.
\item A talk by Jenny Bryan (rstd.io/row-work) inspired the fundamental
implementation details.
}
}

\examples{
library(dplyr)
library(ggplot2)

census <- data.frame(
  quadrat = rep(c("0000", "0001"), each = 3),
  sp = rep(paste0("sp", 1:3), 2),
  n = sample.int(100, 6),
  stringsAsFactors = FALSE
)
census

ungrouped <- census
vgn_diversity(ungrouped, n)

by_quadrat <- group_by(census, quadrat)
vgn_diversity(by_quadrat, n)

vgn_diversity(by_quadrat, n, index = "shannon")

# The output of `vgn_diversity` flows well into common pipelines:

diversity <- census \%>\%
  group_by(quadrat) \%>\%
  vgn_diversity(n)
diversity

# A plot
diversity \%>\%
  ggplot(aes(quadrat, value)) +
  geom_col(aes(fill = index), position = "dodge")

# A summary
diversity \%>\%
  group_by(index) \%>\%
  summarise(mean = mean(value))
}
\seealso{
\code{\link[vegan:diversity]{vegan::diversity()}}.
}
