% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev-pkgdown_template.R
\name{write_pkgdown_template}
\alias{write_pkgdown_template}
\title{Internal funciton to output the template for the package's website}
\usage{
write_pkgdown_template(dir = "./")
}
\arguments{
\item{dir}{Directory where the file \code{pkgdown.yml} will be placed.}
}
\value{
Outputs the file \code{_pkgdown.yml} (in working directory by default).
}
\description{
Internal funciton to output the template for package website built with
the \strong{pkgdown} package (http://hadley.github.io/pkgdown/).
}
\details{
Does not overwrite existing file.
}
\examples{
\dontrun{
write_pkgdown_template()
}

}
\seealso{
\code{\link[pkgdown:build_site]{pkgdown::build_site()}}
}
\keyword{internal}
