% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{abundance}
\alias{abundance}
\title{Abundance.}
\usage{
abundance(x, ...)
}
\description{
\itemize{
\item Copy of \code{\link[dplyr:count]{dplyr::count()}}.
\item Supports groups.
\item Returns 0 when fed with a 0-row dataframe.
}
}
\details{
\code{\link[fgeo.abundance:abundance]{fgeo.abundance::abundance()}} versus ctfs::abundance():
\itemize{
\item When fed with a 0-row dataframe,
\itemize{
\item \code{ctfs::abundance} returns a dataframe with zero rows and zero columns.
\item \code{\link[fgeo.abundance:abundance]{fgeo.abundance::abundance()}} returns a dataframe column \code{n = 0}.
}
\item \code{ctfs::abundance} handles only two grouping variables, via arguments
\code{split1} and \code{split2}.
\item \code{\link[fgeo.abundance:abundance]{fgeo.abundance::abundance()}} handles any number of grouping variables via
\code{\link[dplyr:group_by]{dplyr::group_by()}}.
\item \code{ctfs::abundance} returns a list, which is difficult to visualize and use.
\item \code{\link[fgeo.abundance:abundance]{fgeo.abundance::abundance()}} returns a dataframe subclass tibble, which
prints well even with large datasets and works well with a larger number of
tools for data analysis (e.g. \strong{ggplot2} and \strong{dplyr}).
}
}
\section{Flags}{

Warns if it detects:
\itemize{
\item duplicated values of treeis.
\item multiple values of censusid.
\item multiple values of plotname.
}
}

\section{Warning}{

To pick specific rows (e.g. to pick "alive" stems or \code{dbh} within some range)
you should use functions such as \code{[}, \code{\link[=subset]{subset()}}, or \code{\link[dplyr:filter]{dplyr::filter()}}
\emph{before} using \code{abundance()}. (Notice the difference with \code{ctfs::abundance},
which includes arguments to deal with \code{dbh} and \code{status}).
}

