% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byyr.R
\name{abundance_byyr}
\alias{abundance_byyr}
\alias{basal_area_byyr}
\title{Create tables of abundance and basal area by (round mean) year.}
\usage{
abundance_byyr(vft, ...)

basal_area_byyr(vft, ...)
}
\arguments{
\item{vft}{A ForestGEO-like dataframe; particularly a ViewFullTable.}

\item{...}{An expression of dbh to pick woods of a specific range.}
}
\value{
A dataframe.
}
\description{
Create tables of abundance and basal area by (round mean) year.
}
\examples{
library(fgeo.tool)

vft <- example_byyr
vft
abundance_byyr(vft, DBH >= 10, DBH < 20)
abundance_byyr(vft, DBH >= 10)
basal <- basal_area_byyr(vft, DBH >= 10)
# Convert units and standardize by plot size in hectares
years <- c("2001", "2002")
in_he <- convert_unit_at(basal, .at = years, from = "mm2", to = "hectare")
standardize_at(in_he, .at = years, denominator = 50)
}
