% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_.R
\name{abundance_n}
\alias{abundance_n}
\title{Count individuals}
\usage{
abundance_n(x, group_by = c("quadrat", "sp"), only.alive = TRUE)
}
\arguments{
\item{x}{Census.}

\item{group_by}{Sting giving the variables to group by.}

\item{only.alive}{}
}
\description{
Count individuals
}
\examples{
stem <- bci12t7mini
(n <- abundance_n(stem))
class(n)

abundance_n(stem, group_by = "sp", F)
}
