% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_alive.R
\name{add_id}
\alias{add_id}
\title{Add unique identifier combining treeID and stemID.}
\usage{
add_id(x)
}
\arguments{
\item{x}{Census.}
}
\value{
Returns \code{x} with an additional variable \code{id}.
}
\description{
Add unique identifier combining treeID and stemID.
}
