% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{basal_area}
\alias{basal_area}
\alias{ba}
\alias{ba_sum}
\alias{ba_sum}
\title{Basal area of trees.}
\usage{
ba(dbh, dbhunit = "mm")

ba_sum(dbh, dbhunit = "mm", mindbh = 10)
}
\arguments{
\item{dbh}{Stem diameter, usually at breast height (see data dictionary at 
http://ctfs.si.edu/Public/DataDict/data_dict.php).}

\item{dbhunit}{'cm' or 'mm', only used for basal area.}

\item{mindbh}{The minimum diameter above which the counts are done. Trees
smaller than `mindbh` are excluded. If `NULL`, all living trees are
included.}
}
\value{
\itemize{
\item \code{ba()}: A vector of basal area values of same length as the submitted
vector of dbhs.
\item \code{ba_sum()}: A number giving the the basal area summed over all submitted
dbhs.
}
}
\description{
\itemize{
\item \code{ba()}: calculates the individual basal areas in square meters for all
\code{dbh}-values submitted.
\item \code{ba_sum()}: calculates the basal area summed over all submitted \code{dbh} values
(after removing NAs).
}
}
\examples{
# Takes NAs and throws informative warning
(dbh <- c(seq(5, 50, by = 5), NA))
ba(dbh, dbhunit = "cm")  # default dbhunit is milimeters ("mm")

# (From now on, suppressing warnings to avoid distraction)

# Filter stems at or over a minimum dbh and sum them, all in one step
suppressWarnings(
  ba_sum(dbh, mindbh = 30)
)

# Same; a bit longer but may be easier to understand what is going on
over_mindbh <- dbh[dbh >= 30]
suppressWarnings(
  sum(ba(over_mindbh), na.rm = TRUE)
)
}
\author{
Richard Condit.
}
