% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basal_area.R
\name{basal_area}
\alias{basal_area}
\alias{basal_area.default}
\alias{basal_area.data.frame}
\alias{basal_area.numeric}
\alias{add_basal_area}
\title{Basal area of each individual.}
\usage{
basal_area(x, ...)

\method{basal_area}{default}(x, ...)

\method{basal_area}{data.frame}(x, dbh = dbh, ...)

\method{basal_area}{numeric}(x, ...)

add_basal_area(x, ..., dbh = dbh)
}
\arguments{
\item{x}{Numeric vector or dataframe giving the individuals' diameter.}

\item{...}{Other arguments passed to methods.}

\item{dbh}{If \code{x} is a dataframe, \code{dbh} is the bare name of the column giving
individual's diameter.}
}
\value{
Each individual calculation is simply the area of a circle given its
diameter; thus the unit of the returned value is the square of the input
units unit (i.e. for input in mm, output is mm^2; for input in cm, output is
cm^2, and so on).
\itemize{
\item \code{basal_area.numeric()}: A numeric vector giving the basal area of each
individual.
}
}
\description{
\code{basal_area()} is a generic function with methods for numeric vectors and
dataframes. To calculate basal area within groups, feed
\code{basal_area()} with a dataframe previously grouped with \code{group_by()} (see
examples and section See also). \code{group_by()} comes from \strong{dplyr} but for
convenience it is reexported by \strong{fgeo.abundance}. \code{add_basal_area()} adds
a column \code{basal_area}
}
\examples{
library(dplyr)

stem <- tibble::tibble(
  stemID = 1:6,
  quadrat = paste0("000", rep(1:2, each = 3)),
  sp = c(paste0("sp", c(1, 1, 2)), paste0("sp", c(3, 3, 3))),
  dbh = abs(sample(rnorm(100), 6) * 10)
)
stem
basal_area(stem$dbh)
add_basal_area(stem)

# Sum across all rows
basal_area(stem)

# Sum across rows by groups
basal_area(
  group_by(stem, quadrat, sp)
)
# Same but returns all rows
add_basal_area(stem, quadrat, sp)

# Dealing with missing values
stem2 <- stem
stem2[6, 4] <- NA
stem2

# basal_area() of missing dbh in a vector results in NA
basal_area(stem2$dbh)
# add_basal_area() of missing dbh in a dataframe results in NA
# Same if using add_basal_area()
add_basal_area(stem2)

# But basal_area() of missing dbh in a dataframe ignores NA's
basal_area(stem2)
basal_area(group_by(stem2, quadrat, sp))
# Missing dbh is ignored, and `basal_area` refers to the group not each row.
add_basal_area(stem2, quadrat, sp)
}
\seealso{
\code{\link[dplyr:group_by]{dplyr::group_by()}}.
}
\concept{functions that quote one or more arguments using tidy eval.}
