% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basal_area.R
\name{basal_area}
\alias{basal_area}
\alias{basal_area_ind}
\title{Basal area.}
\usage{
basal_area(x, group_by = c("quadrat", "sp"), only_alive = TRUE)

basal_area_ind(diameter)
}
\arguments{
\item{x}{Census data.}

\item{group_by}{Character string giving the names of variables to group by.}

\item{only_alive}{If TRUE, counts only alive individuals (`status == "A"`).
If FALSE counts individuals of all available statuses.}

\item{diameter}{A vector giving the individuals' diameter (generally, \code{x$dbh}).}
}
\value{
\itemize{
\item \code{basal_area_ind()}: A numeric vector giving the basal area of each
individual in the same units as the input vector. This is simply the
area of a circle given its diameter.
}
}
\description{
Basal area of each individual (\code{basal_area_ind()}).
}
\details{
Units of returned values are the square of the input units. For example, if
you input diameter measured in mm, the output will be an area measured in
mm^2.
}
\examples{
stem <- bciex::bci12s7mini

head(stem$dbh)
head(basal_area_ind(stem$dbh))

head(basal_area(stem))

# Silent
head(suppressMessages(basal_area(stem)))
}
\seealso{
Other grouped summaries: \code{\link{abundance}}
}
