% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byyr.R
\name{byyr}
\alias{byyr}
\alias{byyr_abundance}
\alias{byyr_basal_area}
\title{Create tables of abundance and basal area by (round mean) year.}
\usage{
byyr_abundance(vft)

byyr_basal_area(vft)
}
\arguments{
\item{vft}{A dataframe; particularly a ForestGEO ViewFullTable.}
}
\value{
A dataframe.
}
\description{
Create tables of abundance and basal area by (round mean) year.
}
\examples{
# Or load all fgeo packages in one step with library(fgeo)
library(fgeo.tool)
library(fgeo.base)
  
# FIXME: ADD DEAD TREE FOR CENSUS 2
# There are two trees of the same species in census 1 and one tree in census
# 2, each tree has two stems.
vft_toy <- tibble(
  PlotName = c("luq", "luq", "luq", "luq", "luq", "luq"),
  Status = c("alive", "dead", "alive", "gone", "alive", "dead"),
  ExactDate = c("2001-01-01", "2001-01-01", "2001-01-01", "2001-01-01",
    "2002-01-01", "2002-01-01"),
  TreeID = c("1", "1", "2", "2", "1", "1"),
  StemID = c("1.1", "1.2", "2.1", "2.2", "1.2", "1.1"),
  Genus = c("A", "A", "A", "A", "A", "A"),
  SpeciesName = c("a", "a", "a", "a", "a", "a"),
  PlotCensusNumber = c("1", "1", "1", "1", "2", "2"),
  CensusID = c("1", "1", "1", "1", "2", "2"),
  DBH = c(1L, 6L, 2L, 3L, 5L, 4L),
  Family = c("f", "f", "f", "f", "f", "f"),
  Tag = c("1", "1", "2", "2", "1", "1")
)
vft_toy

byyr_abundance(vft_toy)

ba <- byyr_basal_area(vft_toy)

# Convert units and standardize by plot size in hectares
years <- c("2001", "2002")
in_he <- convert_unit_at(ba, .at = years, from = "mm2", to = "hectare")
standardize_at(in_he, .at = years, denominator = 50)
}
