% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byyr.R
\name{byyr}
\alias{byyr}
\alias{byyr_abundance}
\alias{byyr_basal_area}
\title{Create tables of abundance and basal area by (round mean) year.}
\usage{
byyr_abundance(vft)

byyr_basal_area(vft)
}
\arguments{
\item{vft}{A dataframe; particularly a ForestGEO ViewFullTable.}
}
\value{
A dataframe.
}
\description{
Create tables of abundance and basal area by (round mean) year.
}
\examples{
library(fgeo.base)
library(fgeo.tool)

vft <- data.frame(
  PlotName = "p",
  PlotCensusNumber = c(1, 1, 2, 2, 1, 2),
  CensusID = c(1, 1, 2, 2, 1, 2),
  ExactDate = c(
    "2001-01-01", "2001-01-01", "2002-01-01", "2002-01-01",
    "2001-01-01", "2002-01-01"
  ),
  Tag = c("0001", "0001", "0001", "0001", "0002", "0002"),
  TreeID = c("0001", "0001", "0001", "0001", "0002", "0002"),
  SteemID = c(  "1",    "2",    "1",    "2",    "1",    "1"),
  Status = c("alive", "alive", "alive", "dead", "alive", "dead"),
  DBH = c(11, 15, 12, NA, 21, NA),
  Genus = c("A", "A", "A", "A", "B", "B"),
  SpeciesName = c("a", "a", "a", "a", "b", "b"),
  Family = "f",
  stringsAsFactors = FALSE
)
vft

# First pick the data you want
pick1 <- fgeo.base::pick_plotname(vft, "p")
pick2 <- drop_dead_trees_by_cns(pick1)
pick3 <- pick_dbh_min(pick2, 10)

byyr_abundance(pick3)

ba <- byyr_basal_area(pick3)
ba

# Convert units and standardize by plot size in hectares
years <- c("2000", "2001")
in_he <- convert_unit_at(ba, .at = years, from = "mm2", to = "hectare")
standardize_at(in_he, .at = years, denominator = 50)
}
