% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{count_distinct}
\alias{count_distinct}
\alias{abundance_stem}
\alias{abundance_tree}
\title{Count distinct occurrences of a variable by groups (tree or stem abundance).}
\usage{
count_distinct(.data, .var)

abundance_stem(.data)

abundance_tree(.data)
}
\arguments{
\item{.data}{A dataframe, commonly grouped with \code{group_by()}.}

\item{.var}{A variable to count distinct occurrences.}
}
\value{
An object of the same class as .data. One grouping level will be
dropped.
}
\description{
These functions are usually used on grouped data created by
\code{dplyr::group_by()} (reexported by \strong{fgeo}). The output will have one row
for each group.
\itemize{
\item \code{count_distinct()} counts distinct occurrences of a variable.
\item \code{abundance_stem()} and \code{abundance_tree()} are simpler wrappers for the
specifically counting distinct occurrences of the variables \code{stemID} and
\code{treeID} which uniquely identify each stem and tree in census datasets of
ForestGEO (both stem and tree tables).
}
}
\examples{
library(dplyr)

# Example data
census <- tibble::tibble(
  treeID = c(1, 1, 2, 3, 3, 3),
  stemID = c(1, 2, 3, 4, 5, 6),
  quadrat = paste0("000", rep(1:2, each = 3)),
  sp = c(paste0("sp", c(1, 1, 2)), paste0("sp", c(3, 3, 3))),
  dbh = abs(sample(rnorm(100), 6) * 10)
)
census

count_distinct(census, stemID)

by_quad <- group_by(census, quadrat)
count_distinct(by_quad, stemID)
count_distinct(by_quad, treeID)

abundance_stem(by_quad)
abundance_tree(by_quad)

by_sp <- group_by(census, sp)
abundance_stem(by_sp)
abundance_tree(by_sp)

by_quad_sp <- group_by(census, quadrat, sp)
abundance_stem(by_quad_sp)
abundance_tree(by_quad_sp)
}
\seealso{
\code{\link[dplyr:group_by]{dplyr::group_by()}}, \code{\link[dplyr:summarise]{dplyr::summarise()}}.
}
