% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_woods.R
\name{count_woods}
\alias{count_woods}
\alias{count_trees}
\alias{count_saplings}
\title{Count unique trees based on the dbh of their largest stem.}
\usage{
count_woods(.data, ...)

count_trees(.data)

count_saplings(.data)
}
\arguments{
\item{.data}{A dataframe; particularly a ForestGEO census or ViewFullTable.}

\item{...}{Expressions to pick stems of specific \code{dbh} -- where \emph{stems}
refers to the largest stem of each tree.}
}
\description{
Count unique trees based on the dbh of their largest stem.
}
