% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_woods.R
\name{count_woods}
\alias{count_woods}
\alias{count_trees}
\alias{count_saplings}
\title{Count unique trees based on the dbh of their largest stem.}
\usage{
count_woods(.data, ..., .collapse = collapse_treeid_max)

count_trees(.data, .collapse = collapse_treeid_max)

count_saplings(.data, .collapse = collapse_treeid_max)
}
\arguments{
\item{.data}{A dataframe; particularly a ForestGEO census or ViewFullTable.}

\item{...}{Expressions to pick stems of specific \code{dbh} -- where \emph{stems}
refers to the largest stem of each tree.}
}
\description{
This functions help you to count unique treeid in each census:
\itemize{
\item \code{count_woods()} is general and allows you to filter the data using any
expressions passed via the argument \code{...} to \code{dplyr::filter()}.
\item \code{count_trees()} is an opinionated shortcut that defines tress as stems of
dbh of 100 mm and above.
\item \code{count_samplings()} is an opinionated shortcut that defines saplings as
stems of dbh of between 10 mm inlcusive and 100 mm exclusive.
}
}
