% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{drp}
\alias{drp}
\title{Internal.}
\usage{
drp(x)
}
\description{
Internal.
}
\seealso{
Other functions from http://ctfs.si.edu/Public/CTFSRPackage/: \code{\link{calcMortIndivTable}},
  \code{\link{demography}}, \code{\link{fill.dimension}},
  \code{\link{find.climits}},
  \code{\link{mortality.calculation}},
  \code{\link{mortality.dbh}},
  \code{\link{mortality.eachspp}}
}
\keyword{internal}
