% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{drp}
\alias{drp}
\title{Internal.}
\usage{
drp(x)
}
\description{
Internal.
}
\seealso{
Other functions from http://ctfs.si.edu/Public/CTFSRPackage/: \code{\link{fill.dimension}},
  \code{\link{find.climits}}
}
\keyword{internal}
