% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctfs-recruitment.R
\name{fill.dimension}
\alias{fill.dimension}
\title{Internal.}
\usage{
\method{fill}{dimension}(dataarray, class1, class2, fill = 0)
}
\description{
Internal.
}
\seealso{
Other functions from http://ctfs.si.edu/Public/CTFSRPackage/: \code{\link{drp}},
  \code{\link{find.climits}}, \code{\link{recruitment}}
}
\keyword{internal}
