% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctfs-recruitment.R
\name{find.climits}
\alias{find.climits}
\title{Internal.}
\usage{
find.climits(N, D, alpha = 0.05, kind = "upper")
}
\description{
Internal.
}
\seealso{
Other functions from http://ctfs.si.edu/Public/CTFSRPackage/: \code{\link{drp}},
  \code{\link{fill.dimension}}, \code{\link{recruitment}}
}
\keyword{internal}
