% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{growth}
\alias{growth}
\title{Annual growth rates accross entire census or within groups.}
\usage{
growth(census1, census2, rounddown = FALSE, method = "I", stdev = FALSE,
  dbhunit = "mm", mindbh = 10, growthcol = "dbh", err.limit = 4,
  maxgrow = 75, split1 = NULL, split2 = NULL)
}
\arguments{
\item{census1}{The R Analytical Table for a single census, either tree or
stem.}

\item{census2}{The matching R Analytical Table for a later census.}

\item{rounddown}{If TRUE, all dbh < 55 are rounded down to the nearest
multiple of 5.}

\item{method}{Use 'I' to calculate annual dbh increment: (dbh2 - dbh1)/time,
or 'E' to calculate the relative growth rate (log(dbh2) - log(dbh1))/time.}

\item{stdev}{Logical. Default (FALSE) returns confidence limits, otherwise
returns the SD in growth rate per group.}

\item{dbhunit}{'cm' or 'mm'.}

\item{mindbh}{The minimum diameter above which the counts are done. Trees
smaller than `mindbh` are excluded. If `NULL`, all living trees are
included.}

\item{growthcol}{defines how growth is measured, either 'dbh'or
'agb'(agb=biomass)}

\item{err.limit}{A number. Numbers such as 10000 are high and will return all
measures.}

\item{maxgrow}{A number. Numbers such as 10000 are high and will return all
measures.}

\item{split1}{a vector of categories, one per individual}

\item{split2}{another vector of categories, one per individual}
}
\value{
\code{growth()} retruns a list with components:
\itemize{
\item \code{rate}, the mean annualized growth rate per category selected, either dbh
increment, or relative growth
\item \code{N}, the number of individuals included in the mean (not counting any
excluded)
\item \code{clim}, width of confidence interval; add this number to the mean rate to
get upper confidence limit, substract to get lower
\item \code{dbhmean}, mean dbh in census 1 of individuals included
\item \code{time}, mean time interval in years
\item \code{date1}, mean date included individuals were measured in census 1, as
julian object (R displays as date, but treats as integer)
\item \code{date2}, mean date in census 2.
}
}
\description{
Annual growth rates accross entire census or within groups.
}
\details{
\code{growth()} also requires the columns \code{status} and \code{stemID} in order to
determine which stems should have dbh change calculated.
}
\examples{
# Filtering only 200 trees to make the data small
few_trees <- sample(unique(bci12s6mini$tag), 200)
census1 <- subset(bci12s6mini, tag \%in\% few_trees)
census2 <- subset(bci12s7mini, tag \%in\% few_trees)

growth(census1, census2)

# View just a few species with `head()`
lapply(growth(census1, census2, split1 = census1$sp), head)

split_by_two <- growth(census1, census2, 
  split1 = census1$sp, split2 = census1$quadrat)
lapply(split_by_two, function(x) head(x[1:6]))
str(split_by_two)

}
\seealso{
Other functions from http://ctfs.si.edu/Public/CTFSRPackage/: \code{\link{CI}},
  \code{\link{IfElse}}, \code{\link{addBinParam}},
  \code{\link{arrangeParam.Gibbs}},
  \code{\link{arrangeParam.llike}},
  \code{\link{bad.binparam}},
  \code{\link{binGraphManySpecies.Panel}},
  \code{\link{binGraphManySpecies}},
  \code{\link{binGraphSampleSpecies}},
  \code{\link{biomass.growth}},
  \code{\link{calcMortIndivTable}},
  \code{\link{calculateBinModel.AIC}},
  \code{\link{calculateBinModel.BIC}},
  \code{\link{colMedians}},
  \code{\link{compare.growthbinmodel}},
  \code{\link{defineBinBreaks}}, \code{\link{defineSDpar}},
  \code{\link{demography}}, \code{\link{dgammadexp}},
  \code{\link{drp}}, \code{\link{enoughSamplePerBin}},
  \code{\link{extract.growthdata}},
  \code{\link{fill.dimension}}, \code{\link{find.climits}},
  \code{\link{graph.growthmodel.spp}},
  \code{\link{graph.growthmodel}},
  \code{\link{graph.outliers.spp}},
  \code{\link{graph.outliers}},
  \code{\link{growth.biomass.indiv}},
  \code{\link{growth.dbh}}, \code{\link{growth.eachspp}},
  \code{\link{growth.flexbin}}, \code{\link{growth.indiv}},
  \code{\link{individual_grow.table}},
  \code{\link{linear.model.ctr}},
  \code{\link{linear.model}},
  \code{\link{linearmodel.bin.set}},
  \code{\link{linearmodel.bin}},
  \code{\link{llike.GaussModelSD}},
  \code{\link{llike.GaussModel}},
  \code{\link{llike.linearbin.optim}}, \code{\link{logit}},
  \code{\link{metrop1step}}, \code{\link{model.xy}},
  \code{\link{mortality.calculation}},
  \code{\link{mortality.dbh}},
  \code{\link{mortality.eachspp}},
  \code{\link{overlay.growthbinmodel}}, \code{\link{pst}},
  \code{\link{pts.to.interceptslope}},
  \code{\link{rndown5}},
  \code{\link{run.growthbin.manyspp}},
  \code{\link{run.growthfit.bin}},
  \code{\link{trim.growth}}, \code{\link{wideEnoughBins}}

Other functions to compare two censuses: \code{\link{demography}}
}
