% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_sp.R
\name{map_sp}
\alias{map_sp}
\alias{map_sp_pdf}
\title{Map the distribution of one, some or all species in a census data set.}
\usage{
map_sp(census, species, ...)

map_sp_pdf(census, species, file = "map.pdf", ...)
}
\arguments{
\item{census}{Census data.}

\item{species}{A string of the species codes to plot (\code{sp}).}

\item{...}{Arguments passed to \code{\link[ggplot2:geom_layer]{ggplot2::geom_layer()}} via
\code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.}

\item{file}{A character string giving the name of the file.}
}
\value{
A list of plots is always returned:
\itemize{
\item \code{map_sp()} returns it visibly;
\item \code{map_sp_pdf()} returns it invisibly so it can be reused, but its main
}
}
\description{
Map the distribution of one, some or all species in a census data set.
}
\section{Acknowledgement}{

Thanks to Gabriel Arellano and David Kenfack for ideas and feedback.
}

\examples{
\dontrun{
library(dplyr)  # to make it easer to manipulate data
library(gridExtra)  # to arrange multiple plots in one page

# Data for example
census <- forestr::bci12t7mini

# Plot to screen
map_sp(census, "hybapr")

# Tweak plot with arguments passed to [ggplot2::layer()]
map_sp(census, c("hybapr", "faraoc"), color = "blue", size = 3, shape = 1)

# Selecting only a few abundant species
count_sp <- dplyr::count(census, sp)
decreasing_n <- dplyr::arrange(count_sp, desc(n))
top_n <- decreasing_n$sp[1:8]

# Plot to screen
one_per_page <- map_sp(census, top_n)
one_per_page

# Print multiple page in a single .pdf file.
# Invisibly returning a list of plots that can be reused
one_per_page <- map_sp_pdf(census, top_n)

# To export .png wrap your plots between png() and dev.off()
png()
one_per_page[1:3]
dev.off()

# Reusing one_per_page
multi_per_page <- gridExtra::marrangeGrob(one_per_page, ncol = 2, nrow = 2)
multi_per_page

# Saving the multipage into a .pdf
pdf("multi_paged.pdf", paper = "letter")
multi_per_page
dev.off()
}
}
\seealso{
\code{\link[ggplot2:geom_layer]{ggplot2::geom_layer()}}, \code{\link[grDevices:pdf]{grDevices::pdf()}}, \code{\link[grDevices:png]{grDevices::png()}}.
}
