% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctfs-recruitment.R
\name{recruitment}
\alias{recruitment}
\alias{recruitment.eachspp}
\alias{recruitment_df}
\alias{recruitment.eachspp_df}
\title{Caculate recruitment for all and each species (as in the CTFS R package)}
\usage{
recruitment(census1, census2, mindbh = 10, alivecode = c("A", "AB", "AS"),
  split1 = NULL, split2 = NULL)

recruitment.eachspp(census1, census2, mindbh = 10, alivecode = c("A", "AB",
  "AS"))

recruitment_df(...)

recruitment.eachspp_df(...)
}
\arguments{
\item{census1}{The R Analytical Table for a single census, either tree or
stem.}

\item{census2}{The matching R Analytical Table for a later census.}

\item{mindbh}{The minimum diameter above which the counts are done. Trees
smaller than `mindbh` are excluded. If `NULL`, all living trees are
included.}

\item{alivecode}{character, codes of the variable `status` that indicate the 
tree is alive. The default 'A' is the standard CTFS designation for living 
trees or stems}

\item{split1}{a vector of categories, one per individual}

\item{split2}{another vector of categories, one per individual}
}
\value{
(xxx ask documentation) A list with components:
\itemize{
\item \code{N2}:
\item \code{R}:
\item \code{rate}:
\item \code{lower}:
\item \code{upper}:
\item \code{time}:
\item \code{date1}:
\item \code{date2}:
}
}
\description{
\code{recruitment()} and \code{recruitment.eachspp()} caculate recruitment for all and
each species and return a list (as in
(http://ctfs.si.edu/Public/CTFSRPackage/). \code{recruitment_df()} and
\code{recruitment.eachspp_df()} are equivalent but instead of a list they output a
dataframe, which is easier to manipulate and view.
}
\details{
For every individual of all species you must provide two complete datasets,
one per census, with variables \code{dbh}, \code{pom}, \code{status} and \code{date}. Any status
indicating a live tree can be submitted in the variable \code{alivecode}.
Survivors are all individuals alive in both censuses, with \code{status == A} in
the first census, and larger than the minimum dbh in the first census. The
total population in the second census includes all those alive, above the
minimum dbh, plus any other survivors. As in \code{\link[=mortality]{mortality()}}, individuals whose
status is NA in either census are deleted from all calculations.
}
\examples{
# Some data to play with
census1 <- forestr::bci12t6mini
census2 <- forestr::bci12t7mini

# A bit ackward to manipulate and view  -----------------------------------

recruitment(census1, census2)

# Trick to fit a few species on screen
each_spp <- recruitment.eachspp(census1, census2)
lapply(each_spp, head)

# Easier to manipulate and view -------------------------------------------

library(dplyr)
library(tidyr)

(result <- recruitment.eachspp_df(census1, census2))

sp_to_explore <- c("casesy", "ocotob", "soroaf")
long_format <- dplyr::filter(result, sp \%in\% sp_to_explore)
long_format

wide_format <- tidyr::spread(long_format, key = sp, value = value)
wide_format
}
\seealso{
Other functions from http://ctfs.si.edu/Public/CTFSRPackage/: \code{\link{drp}},
  \code{\link{fill.dimension}}, \code{\link{find.climits}}
}
