% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{reexports}
\alias{reexports}
\alias{tally}
\alias{count}
\alias{add_tally}
\alias{add_count}
\alias{group_by}
\alias{ungroup}
\title{Reexports from dplyr.}
\description{
Reexports from dplyr.
}
\keyword{internal}
