% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tally}
\alias{reexports}
\alias{add_tally}
\alias{reexports}
\alias{count}
\alias{reexports}
\alias{add_count}
\alias{reexports}
\alias{group_by}
\alias{reexports}
\alias{ungroup}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{tally}}, \code{\link[dplyr]{add_tally}}, \code{\link[dplyr]{count}}, \code{\link[dplyr]{add_count}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{ungroup}}}
}}

