% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgn_diversity.R
\name{vgn_diversity}
\alias{vgn_diversity}
\title{Shannon and Simpson diversity indices.}
\usage{
vgn_diversity(x, abundance, index = c("specnumber", "shannon"))
}
\arguments{
\item{x}{A dataframe.}

\item{abundance}{A numeric column in \code{x} giving the abundance by species.}

\item{index}{Some of c("specnumber", "shannon", "invsimpson", "simpson")}
}
\value{
A summary of the input dataframe, with the additional column
\code{diversity}.
}
\description{
Adds a column \code{diversity}, of Shannon and Simpson diversity indices
calculated with \code{\link[vegan:diversity]{vegan::diversity()}}. The output is a dataframe, convenient
for workflows with general purpose tools such as \strong{dplyr} and \strong{ggplot2}.
}
\section{Acknowledgements}{

\itemize{
\item David Kenfack inspired the need for this function.
\item A talk by Jenny Bryan (rstd.io/row-work) inspired the fundamental
implementation details.
}
}

\examples{
library(dplyr)
library(ggplot2)

census <- data.frame(
  quadrat = rep(c("0000", "0001"), each = 3),
  sp = paste0("sp", c(1:3, 1, 4, 5)),
  n = sample.int(100, 6),
  stringsAsFactors = FALSE
)
census <- census[1:5, ]
census

ungrouped <- census
vgn_diversity(ungrouped, n)

by_quadrat <- group_by(census, quadrat)
vgn_diversity(by_quadrat, n)
# Is a shortcut for ...

div <- mutate(
  by_quadrat,
  specnumber = vegan::specnumber(n),
  shannon = vegan::diversity(n)
)
gather(ungroup(div), "index", "value", specnumber:shannon) \%>\% 
  select(quadrat, index, value) \%>\% 
  unique() \%>\% 
  arrange(quadrat, desc(index))

# The output of `vgn_diversity` flows well into common pipelines:

diversity <- census \%>\%
  group_by(quadrat) \%>\%
  vgn_diversity(n)
diversity

# A plot
diversity \%>\%
  ggplot(aes(quadrat, value)) +
  geom_col(aes(fill = index), position = "dodge")

# A summary
diversity \%>\%
  group_by(index) \%>\%
  summarise(mean = mean(value))
}
\seealso{
\code{\link[vegan:diversity]{vegan::diversity()}}.
}
